ffmpeg and libav compatibility fixes, round two.
Change how the delay is calculated (so it's actually correct).  It now relies
on the ffmpeg or libav version numbers to determine the appropriate method it
needs to use.  The calculation of the delay is now exactly the same as the way
it was previously done internally in libav/ffmpeg.

git-svn-id: http://ffmpegsource.googlecode.com/svn/trunk@595 dcf88d3e-5074-11de-ac45-ff9ca5716a9b