Fix handling of audio delay with invalid inital timestamps
When the first audio PTS in a file is AV_NOPTS_VALUE, it was overflowing
int64_t, resulting in the delay being 0 with MSVC, and possibly other
values with other compiles. To fix this, use the first valid PTS and
calculate the elapsed time before that packet from the sample counts.

git-svn-id: http://ffmpegsource.googlecode.com/svn/trunk@700 dcf88d3e-5074-11de-ac45-ff9ca5716a9b